refactor: fix sonarqube accessibility and code quality issues
- replace non-interactive div[role=button] with native <button>
- extract nested component definitions to module level
- replace nested ternaries with early-return helper functions
- mark component props as readonly
- prefer globalThis over window, Number.parseInt, Set, optional chaining
- remove unused destructuring and unnecessary type assertions
- add missing iframe title and specify eslint-disable rule
