+++ public/js/pathname-init.js
@@ -2,8 +2,8 @@
   try {
     sessionStorage.setItem(
       'konsulin_initial_pathname',
-      window.location.pathname
-    )
-    sessionStorage.removeItem('konsulin_reload_anonymous_done')
-  } catch (e) {} // eslint-disable-line
-})()
+      globalThis.location.pathname
+    );
+    sessionStorage.removeItem('konsulin_reload_anonymous_done');
+  } catch {}
+})();
