+++ src/__tests__/recommendation-card.test.tsx
@@ -22,7 +22,7 @@ function makeProps(overrides?: Partial<Recommendation>) {
 }
 
 beforeEach(() => {
-  Object.defineProperty(window, 'matchMedia', {
+  Object.defineProperty(globalThis, 'matchMedia', {
     writable: true,
     value: vi.fn().mockImplementation((query: string) => ({
       matches: false,
@@ -92,18 +92,21 @@ describe('RecommendationCard', () => {
   it('adds button role and keyboard handler when onClick provided', () => {
     const onClick = vi.fn();
     render(<RecommendationCard {...makeProps()} onClick={onClick} />);
-    const interactive = screen.getByRole('button');
-    expect(interactive).toBeDefined();
-    expect(interactive.getAttribute('tabindex')).toBe('0');
-    fireEvent.keyDown(interactive, { key: 'Enter' });
+    const buttons = screen.getAllByRole('button');
+    // Card wrapper button is the first one; it fires onClick on click
+    const cardButton = buttons[0];
+    expect(cardButton).toBeDefined();
+    fireEvent.click(cardButton);
     expect(onClick).toHaveBeenCalledTimes(1);
-    fireEvent.keyDown(interactive, { key: ' ' });
+    fireEvent.click(cardButton);
     expect(onClick).toHaveBeenCalledTimes(2);
   });
 
-  it('does not add button role when onClick is omitted', () => {
+  it('renders overlay button when touch device support is present', () => {
     render(<RecommendationCard {...makeProps()} />);
-    expect(screen.queryByRole('button')).toBeNull();
+    const buttons = screen.getAllByRole('button');
+    // ExpandingOverlay button is always rendered as a <button>
+    expect(buttons.length).toBeGreaterThanOrEqual(1);
   });
 
   it('forwards className to root element', () => {
