+++ src/app/home-content-clinician.tsx
@@ -9,6 +9,42 @@ import { Calendar, Dumbbell, FileText } from 'lucide-react';
 import Link from 'next/link';
 import { useMemo } from 'react';
 
+type SessionRowData = {
+  slotStart?: string;
+  slotEnd?: string;
+  appointmentId?: string;
+  displayPatientName: string;
+};
+
+function SessionRow({ session }: Readonly<{ session: SessionRowData }>) {
+  if (!session.slotStart || !session.slotEnd) return null;
+  const startTime = format(parseISO(session.slotStart), 'HH:mm');
+  const endTime = format(parseISO(session.slotEnd), 'HH:mm');
+  const isPast = parseISO(session.slotEnd).getTime() < Date.now();
+
+  return (
+    <div className='flex items-center gap-3 px-4 py-3'>
+      <div className='min-w-[60px] text-center'>
+        <div className='text-[13px] font-bold text-gray-800'>{startTime}</div>
+        <div className='text-[10px] text-gray-400'>{endTime}</div>
+      </div>
+      <div
+        className={`h-8 w-[3px] rounded-full ${isPast ? 'bg-gray-200' : 'bg-[#13C2C2]'}`}
+      />
+      <div className='flex-1'>
+        <div className='text-[12px] font-bold text-gray-800'>
+          {session.displayPatientName}
+        </div>
+      </div>
+      <div
+        className={`rounded-full px-2 py-0.5 text-[10px] ${isPast ? 'bg-gray-100 text-gray-400' : 'bg-[#E6F7F7] text-[#13C2C2]'}`}
+      >
+        {isPast ? 'Completed' : 'Upcoming'}
+      </div>
+    </div>
+  );
+}
+
 /** Clinician home page showing today's schedule and quick actions. */
 export default function HomeContentClinician() {
   const { state: authState, isLoading: isAuthLoading } = useAuth();
@@ -57,42 +93,6 @@ export default function HomeContentClinician() {
         <div className='flex gap-4'>
           <Skeleton className='h-[80px] w-full rounded-lg bg-[hsl(210,40%,96.1%)]' />
           <Skeleton className='h-[80px] w-full rounded-lg bg-[hsl(210,40%,96.1%)]' />
-        </div>
-      </div>
-    );
-  }
-
-  type SessionRowData = {
-    slotStart?: string;
-    slotEnd?: string;
-    appointmentId?: string;
-    displayPatientName: string;
-  };
-
-  function SessionRow({ session }: Readonly<{ session: SessionRowData }>) {
-    if (!session.slotStart || !session.slotEnd) return null;
-    const startTime = format(parseISO(session.slotStart), 'HH:mm');
-    const endTime = format(parseISO(session.slotEnd), 'HH:mm');
-    const isPast = parseISO(session.slotEnd).getTime() < Date.now();
-
-    return (
-      <div className='flex items-center gap-3 px-4 py-3'>
-        <div className='min-w-[60px] text-center'>
-          <div className='text-[13px] font-bold text-gray-800'>{startTime}</div>
-          <div className='text-[10px] text-gray-400'>{endTime}</div>
-        </div>
-        <div
-          className={`h-8 w-[3px] rounded-full ${isPast ? 'bg-gray-200' : 'bg-[#13C2C2]'}`}
-        />
-        <div className='flex-1'>
-          <div className='text-[12px] font-bold text-gray-800'>
-            {session.displayPatientName}
-          </div>
-        </div>
-        <div
-          className={`rounded-full px-2 py-0.5 text-[10px] ${isPast ? 'bg-gray-100 text-gray-400' : 'bg-[#E6F7F7] text-[#13C2C2]'}`}
-        >
-          {isPast ? 'Completed' : 'Upcoming'}
         </div>
       </div>
     );
