+++ src/components/quick-action-fab.tsx
@@ -55,14 +55,10 @@ export default function QuickActionFab() {
   return (
     <>
       {isOpen && (
-        <div
-          role='button'
-          tabIndex={0}
+        <button
+          type='button'
           className='animate-overlay-in fixed inset-0 z-40 bg-black/80'
           onClick={close}
-          onKeyDown={e => {
-            if (e.key === 'Enter' || e.key === ' ') close();
-          }}
           aria-label='Close menu'
         />
       )}
